Wrote script to read and translate CSV files into JSON for bulk insertion to DB. Added new transaction get routes.

+++ backend/src/models/model_transaction.ts
@@ -23,10 +23,10 @@ export const Transaction = sq.define('transaction', {
         allowNull: false
     },
     credit: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.REAL,
     },
     debit: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.REAL,
     },
     description: {
         type: DataTypes.CHAR,
